Only wait one hour for Hexo changes to test before raising the PR.
The `prCreation` property we already have set to `not-pending` causes the PR
not to be raised until the tests either pass or fail to prevent the noise
created by PRs before they're ready to merge.

This works generally great, however repositories which do not have tests on
branches which are merely pushed do not ever turn into PRs because they have
no tests configured and nothing to turn green (or red!).

This sets an upper bound of one hour for how long we'll wait for those
tests, which should be more than enough in most cases.

Ref: https://renovateapp.com/docs/configuration-reference/configuration-options#prnotpendinghours